pyproject.toml
[build-system]
requires = ["setuptools>=65.5", "pip>=22.3"]
build-backend = "setuptools.build_meta"

[project]
name = "invoke-training"
version = "0.0.1"
authors = [{ name = "The Invoke AI Team", email = "ryan@invoke.ai" }]
description = "A library for Stable Diffusion model training."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "accelerate~=0.21.0",
    "datasets~=2.14.3",
    "diffusers~=0.19.3",
    "numpy",
    "omegaconf",
    "Pillow",
    "prodigyopt",
    "pydantic",
    "pyyaml",
    "safetensors",
    "tensorboard",
    "torch>=2.0.1",
    "torchvision~=0.15.2",
    "tqdm",
    "transformers~=4.31.0",
    # Known issue with xformers 0.0.16 on some GPUs:
    # https://github.com/huggingface/diffusers/issues/2234#issuecomment-1416931212
    "xformers>=0.22.0",
    "gradio",
]

[project.optional-dependencies]
"test" = [
    "black~=23.7.0",
    "isort~=5.12.0",
    "pre-commit~=3.3.3",
    "pytest~=7.4.0",
    "ruff~=0.0.281",
]

[project.scripts]
"invoke-finetune-lora-sd" = "invoke_training.scripts.invoke_finetune_lora_sd:main"
"invoke-finetune-lora-sdxl" = "invoke_training.scripts.invoke_finetune_lora_sdxl:main"
"invoke-dreambooth-lora-sd" = "invoke_training.scripts.invoke_dreambooth_lora_sd:main"
"invoke-dreambooth-lora-sdxl" = "invoke_training.scripts.invoke_dreambooth_lora_sdxl:main"
"invoke-generate-images" = "invoke_training.scripts.invoke_generate_images:main"

[project.urls]
"Homepage" = "https://github.com/invoke-ai/invoke-training"
"Discord" = "https://discord.gg/ZmtBAhwWhy"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.ruff]
select = ["E", "F", "W", "C9", "N8"]
target-version = "py39"
line-length = 120

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
    "cuda: marks tests that require a CUDA GPU",
    "loads_model: marks tests that require a model (or data) from the HF hub",
]
